FEAT: Filtrando dados por usuário logado

+++ studies/views.py
@@ -1,7 +1,7 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.http import Http404
-from django.shortcuts import redirect, render
+from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
 from .forms import StudyForm, SubjectForm
@@ -10,8 +10,9 @@ from .models import Study, Subject
 
 @login_required(login_url='accounts:login', redirect_field_name='next')
 def home(request):
-    studies = Study.objects.all().order_by('start_time')
-    subjects = Subject.objects.all()
+    studies = Study.objects.filter(
+        subject__student=request.user).order_by('start_time')
+    subjects = Subject.objects.filter(student=request.user)
 
     studies_form_data = request.session.get('studies_form_data') or None
     subject_form_data = request.session.get('subject_form_data') or None
@@ -56,7 +57,10 @@ def create_subject(request):
     form = SubjectForm(request.POST)
 
     if form.is_valid():
-        form.save()
+        form_subject = form.save(commit=False)
+        form_subject.student = request.user
+        form_subject.save()
+
         del (request.session['subject_form_data'])
         messages.success(request, 'Assunto adicionado')
     else:
@@ -67,7 +71,10 @@ def create_subject(request):
 
 @login_required(login_url='accounts:login', redirect_field_name='next')
 def delete_study(request, id):
-    study = Study.objects.get(id=id)
+    study = get_object_or_404(Study, subject__student=request.user, pk=id)
+
+    if not study:
+        raise Http404()
 
     study.delete()
     messages.info(request, 'Deletado com sucesso!')
@@ -77,7 +84,7 @@ def delete_study(request, id):
 
 @login_required(login_url='accounts:login', redirect_field_name='next')
 def delete_subject(request, id):
-    subject = Subject.objects.get(id=id)
+    subject = get_object_or_404(Subject, student=request.user, pk=id)
 
     subject.delete()
     messages.info(request, 'Deletado com sucesso!')
@@ -87,7 +94,8 @@ def delete_subject(request, id):
 
 @login_required(login_url='accounts:login', redirect_field_name='next')
 def subject(request, id):
-    subject = Subject.objects.get(id=id)
+    subject = get_object_or_404(Subject, student=request.user, pk=id)
+
     return render(request, 'studies/pages/subject.html', context={
         'subject': subject,
     })
